requirements.txt
boto3==1.10.28
botocore==1.13.28
Django==2.2.7
django-crispy-forms==1.8.1
django-static-fontawesome==5.10.2
django-storages==1.8
django-tinymce==2.8.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.4
Pillow==6.2.1
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
